+++ src/main/java/restapi/repository/ContatoRepository.java
@@ -3,6 +3,6 @@ package restapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import restapi.model.Contato;
 
-public interface ContatoRepository extends CrudRepository<Contato, Integer> {
+public interface ContatoRepository extends CrudRepository<Contato, Long> {
 
 }
